New NARC API
So it turns out that NSMB and LoZ:ST use incompatible NARC headers. Thus, there's more state involved with a NARC than just the files list and the filenames table, and it therefore makes sense to write a whole class for it. The benefits include an API that matches that of ROM very closely, and extra convenience functions for retrieving and replacing files by name. The drawback is that the API has no similarity whatsoever with the old one, and any code written for the old API needs to be updated, guaranteed. Also, new tutorial!