Initial unstable support for BNBL and BNCL
BNCD should obviously be supported, too, at some point, but I'll just add this for now. Note: the APIs for alignment stuff could arguably be made more low-level, with named constants or an enum for the various alignment types. Not sure if that's a better approach than this or not.