graphics2D API improvements
Completely backwards-incompatible, but having to special-case all of one's code for both 4bpp and 8bpp was kind of ridiculous. Especially when the nybble order for 4bpp is opposite of what you'd expect (at least in my opinion). This is definitely less efficient, but efficiency isn't really a goal of ndspy -- rather, hiding binary-level details is (such as letting you avoid having to split up 4bpp bytes objects manually).