color.py: Add expand and contract functions
These convert between (0-31, 0-31, 0-31, 0-1) and (0-255, 0-255, 0-255, 0-255) RGBA tuples.

Also made a=1 by default in pack(). The default was originally 0 because most colors are transparent in 2D background layers in most cases (and the hardware ignores that), but having opaque be the default is both less surprising for the user, and more consistent with pack255() and the new expand() function.