New API for ndspy.color
I left the old one in for now, because I'm not sure if I'm going to stick with this new one. It's a lot more flexible, but also about twice as slow, and takes significantly more typing to use. (Also, the code for it is necessarily super ugly, unfortunately.)

I'll have to think about it some more, but for now, I'm committing it anyway since the module is undocumented and I can therefore do that.