Change NintendoDSRom.__repr__ a bit more
Ideally, __repr__() should return a string showing an expression that would evaluate to an equivalent object. I think it's OK to make an exception in this case, since ROMs can be huge, but in that case, it should look visually obvious that it's not valid Python syntax. The class's existing __str__() is handy for that.